requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
Django==2.0.6
djangorestframework==3.8.2
idna==2.7
pycparser==2.18
pyOpenSSL==18.0.0
pytz==2018.4
requests==2.19.1
six==1.11.0
urllib3==1.23
